# server/go/model_domain.go
/*
 * pdns-api Domains
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Domain struct {

	Id int32 `json:"id"`

	Name string `json:"name"`

	Master string `json:"master,omitempty"`

	LastCheck int32 `json:"last_check,omitempty"`

	Type string `json:"type"`

	NotifiedSerial int32 `json:"notified_serial,omitempty"`

	Account string `json:"account,omitempty"`
}